# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tatame360 API
  version: 0.1.0
  description: Contrato da primeira entrega vertical. Novos endpoints devem ser acrescentados antes da implementação.
servers:
  - url: http://localhost:18080/api/v1
paths:
  /auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Login'}
      responses:
        '200': {description: Sessão iniciada}
        '401': {$ref: '#/components/responses/Error'}
  /auth/password/forgot:
    post:
      security: []
      summary: Solicita recuperação sem revelar se o e-mail existe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: {email: {type: string, format: email}}
      responses:
        '200': {description: Solicitação recebida}
        '429': {$ref: '#/components/responses/Error'}
  /auth/password/reset:
    post:
      security: []
      summary: Consome token de uso único e encerra as sessões existentes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: {type: string}
                password: {type: string, minLength: 12, maxLength: 72}
      responses:
        '200': {description: Senha redefinida}
        '410': {$ref: '#/components/responses/Error'}
  /auth/invitations/{token}:
    parameters:
      - {name: token, in: path, required: true, schema: {type: string}}
    get:
      security: []
      responses:
        '200': {description: Dados públicos mínimos do convite}
        '410': {$ref: '#/components/responses/Error'}
  /auth/invitations/{token}/accept:
    parameters:
      - {name: token, in: path, required: true, schema: {type: string}}
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                name: {type: string, maxLength: 120}
                password: {type: string, minLength: 12, maxLength: 72}
      responses:
        '200': {description: Convite aceito}
        '410': {$ref: '#/components/responses/Error'}
  /me:
    get:
      responses:
        '200': {description: Usuário, academias e papéis}
        '401': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/students:
    parameters:
      - {$ref: '#/components/parameters/AcademyId'}
    get:
      parameters:
        - {name: search, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
      responses:
        '200': {description: Página de alunos}
        '404': {$ref: '#/components/responses/Error'}
    post:
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StudentInput'}
      responses:
        '200': {description: Aluno e vínculo criados}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/imports/preview:
    parameters: [{$ref: '#/components/parameters/AcademyId'}]
    post:
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties: {file: {type: string, format: binary}}
      responses:
        '200': {description: Lote temporário com validação por linha}
        '422': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/imports/{importId}/confirm:
    parameters:
      - {$ref: '#/components/parameters/AcademyId'}
      - {name: importId, in: path, required: true, schema: {type: string, format: uuid}}
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [acceptPossibleDuplicates]
              properties:
                acceptPossibleDuplicates: {type: boolean}
                classId: {type: [string, 'null'], format: uuid}
      responses:
        '200': {description: Lote confirmado de forma idempotente}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/team:
    parameters: [{$ref: '#/components/parameters/AcademyId'}]
    get:
      responses: {'200': {description: Integrantes e convites da academia}}
  /academies/{academyId}/team/{userId}:
    parameters:
      - {$ref: '#/components/parameters/AcademyId'}
      - {name: userId, in: path, required: true, schema: {type: string, format: uuid}}
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, active]
              properties:
                role: {type: string, enum: [MANAGER, INSTRUCTOR, FRONT_DESK]}
                active: {type: boolean}
      responses:
        '200': {description: Acesso atualizado}
        '403': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/invitations:
    parameters: [{$ref: '#/components/parameters/AcademyId'}]
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, role]
              properties:
                name: {type: string, maxLength: 120}
                email: {type: string, format: email}
                role: {type: string, enum: [MANAGER, INSTRUCTOR, FRONT_DESK]}
      responses:
        '200': {description: Convite de uso único criado}
        '409': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/invitations/{invitationId}/revoke:
    parameters:
      - {$ref: '#/components/parameters/AcademyId'}
      - {name: invitationId, in: path, required: true, schema: {type: string, format: uuid}}
    post:
      responses:
        '200': {description: Convite revogado}
  /academies/{academyId}/class-templates:
    parameters: [{$ref: '#/components/parameters/AcademyId'}]
    get:
      responses: {'200': {description: Turmas acessíveis}}
    post:
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: Turma criada e sessões futuras geradas}}
  /academies/{academyId}/sessions/{sessionId}/roster:
    get:
      parameters:
        - {$ref: '#/components/parameters/AcademyId'}
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
      responses: {'200': {description: Sessão e alunos elegíveis}}
  /academies/{academyId}/sessions/{sessionId}/attendance/batch:
    post:
      parameters:
        - {$ref: '#/components/parameters/AcademyId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentIds, version]
              properties:
                studentIds: {type: array, maxItems: 500, items: {type: string, format: uuid}}
                version: {type: integer, minimum: 0}
      responses:
        '200': {description: Chamada concluída}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /academies/{academyId}/retention/attention:
    get:
      parameters: [{$ref: '#/components/parameters/AcademyId'}]
      responses: {'200': {description: Acompanhamentos abertos}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    AcademyId: {name: academyId, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, format: uuid}}
  responses:
    Error:
      description: Erro padronizado
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Login:
      type: object
      required: [email, password, mobile]
      properties:
        email: {type: string, format: email}
        password: {type: string, maxLength: 72}
        mobile: {type: boolean}
    StudentInput:
      type: object
      required: [name, startsOn]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        email: {type: [string, 'null'], format: email}
        phone: {type: [string, 'null'], maxLength: 30}
        startsOn: {type: string, format: date}
        classId: {type: [string, 'null'], format: uuid}
    Error:
      type: object
      required: [code, message, fieldErrors, requestId]
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors: {type: array}
        requestId: {type: string, format: uuid}
security:
  - bearerAuth: []